requirements.txt
Django==1.11
djangorestframework==3.6.3
algoliasearch-django
xmltodict
psycopg2
django-filter

ipdb
ipython
